Refactored to smart/dumb components
