+++ src/pages/screens/BibliographyListScreen.css
